@@ -14,7 +14,6 @@ classifiers = [
   'Programming Language :: Python :: 3.10',
   'Programming Language :: Python :: 3.11',
   'Programming Language :: Python :: 3.12',
-  'Programming Language :: Python :: 3.9',
 ]
 dependencies = [
   'mne-lsl',
@@ -35,7 +34,7 @@ maintainers = [
 ]
 name = 'lsl_viewpoint'
 readme = 'README.md'
-requires-python = '>=3.9'
+requires-python = '>=3.10'
 version = '0.3.0.dev0'
 
 [project.optional-dependencies]
@@ -52,11 +51,10 @@ full = [
   'lsl_viewpoint[all]',
 ]
 style = [
-  'black',
-  'codespell',
+  'codespell[toml]>=2.2.4',
   'isort',
   'pydocstyle[toml]',
-  'ruff',
+  'ruff>=0.1.8',
   'toml-sort',
   'yamllint',
 ]
@@ -76,18 +74,11 @@ homepage = 'https://github.com/fcbg-hnp-meeg/lsl-viewpoint'
 source = 'https://github.com/fcbg-hnp-meeg/lsl-viewpoint'
 tracker = 'https://github.com/fcbg-hnp-meeg/lsl-viewpoint/issues'
 
-[tool.black]
-extend-exclude = '''
-(
-      __pycache__
-    | .github
-    | .ruff_cache
-    | setup.py
-)
-'''
-include = '\.pyi?$'
-line-length = 88
-target-version = ['py39']
+[tool.codespell]
+check-filenames = true
+check-hidden = true
+ignore-words = '.codespellignore'
+skip = 'build,.git,.mypy_cache,.pytest_cache'
 
 [tool.coverage.report]
 exclude_lines = [
@@ -115,7 +106,7 @@ extend_skip_glob = [
 line_length = 88
 multi_line_output = 3
 profile = 'black'
-py_version = 39
+py_version = 310
 
 [tool.pydocstyle]
 add_ignore = 'D100,D104,D107'
@@ -126,7 +117,7 @@ match-dir = '^lsl_viewpoint.*'
 
 [tool.pytest.ini_options]
 addopts = '--durations 20 --junit-xml=junit-results.xml --verbose'
-minversion = '6.0'
+minversion = '8.0'
 
 [tool.ruff]
 extend-exclude = [
@@ -135,8 +126,19 @@ extend-exclude = [
 ]
 line-length = 88
 
-[tool.ruff.per-file-ignores]
-"__init__.py" = ["F401"]
+[tool.ruff.format]
+docstring-code-format = true
+
+[tool.ruff.lint]
+ignore = []
+select = ["A", "B", "E", "F", "UP", "W"]
+
+[tool.ruff.lint.per-file-ignores]
+'*' = [
+  "B904", # 'Within an except clause, raise exceptions with raise ... from ...'
+]
+'*.pyi' = ['E501']
+'__init__.py' = ['F401']
 
 [tool.setuptools]
 include-package-data = false
