@@ -3,11 +3,17 @@
 Inspired from pandas: https://pandas.pydata.org/
 """
 
+from __future__ import annotations  # c.f. PEP 563, PEP 649
+
 import importlib
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from types import ModuleType
 
 # A mapping from import name to package name (on PyPI) when the package name
 # is different.
-INSTALL_MAPPING = {
+_INSTALL_MAPPING: dict[str, str] = {
     "codespell_lib": "codespell",
     "cv2": "opencv-python",
     "parallel": "pyparallel",
@@ -22,7 +28,7 @@ def import_optional_dependency(
     name: str,
     extra: str = "",
     raise_error: bool = True,
-):
+) -> ModuleType | None:
     """Import an optional dependency.
 
     By default, if a dependency is missing an ImportError with a nice message will be
@@ -41,11 +47,11 @@ def import_optional_dependency(
 
     Returns
     -------
-    module : Optional[ModuleType]
+    module : Module | None
         The imported module when found.
         None is returned when the package is not found and raise_error is False.
     """
-    package_name = INSTALL_MAPPING.get(name)
+    package_name = _INSTALL_MAPPING.get(name)
     install_name = package_name if package_name is not None else name
 
     try:
