@@ -1,8 +1,9 @@
 import platform
 import sys
+from collections.abc import Callable
 from functools import partial
 from importlib.metadata import requires, version
-from typing import IO, Callable, List, Optional
+from typing import IO
 
 import psutil
 from packaging.requirements import Requirement
@@ -10,7 +11,7 @@ from packaging.requirements import Requirement
 from ._checks import check_type
 
 
-def sys_info(fid: Optional[IO] = None, developer: bool = False):
+def sys_info(fid: IO | None = None, developer: bool = False):
     """Print the system information for debugging.
 
     Parameters
@@ -71,14 +72,14 @@ def sys_info(fid: Optional[IO] = None, developer: bool = False):
 
 
 def _list_dependencies_info(
-    out: Callable, ljust: int, package: str, dependencies: List[Requirement]
+    out: Callable, ljust: int, package: str, dependencies: list[Requirement]
 ):
     """List dependencies names and versions."""
     unicode = sys.stdout.encoding.lower().startswith("utf")
     if unicode:
         ljust += 1
 
-    not_found: List[Requirement] = list()
+    not_found: list[Requirement] = list()
     for dep in dependencies:
         if dep.name == package:
             continue
@@ -110,9 +111,11 @@ def _list_dependencies_info(
 
     if len(not_found) != 0:
         not_found = [
-            f"{dep.name} ({str(dep.specifier)})"
-            if len(dep.specifier) != 0
-            else dep.name
+            (
+                f"{dep.name} ({str(dep.specifier)})"
+                if len(dep.specifier) != 0
+                else dep.name
+            )
             for dep in not_found
         ]
         if unicode:
